refactor: extract NAT address/scope/pool helpers (#4846)
Move the shared address-family, host-mask, zone/scope-parsing, pool
address-expansion, and deterministic-NAT helpers out of the 2,578-LOC
compiler_nat.go into a new compiler_nat_helpers.go in the same package.

Pure code motion: each declaration (natAddrFamily, natCIDRIPPart,
isHostMaskAddress, parseZoneList, the natMatchScope type + natScopeKinds
var, parseNATMatchScopes, collectNATScopes, applyNATFromScope,
applyNATToScope, appendPoolAddresses, expandAddressRange, and the three
applyDeterministic* helpers) is moved byte-for-byte; only the import
blocks are adjusted (compiler_nat.go no longer needs encoding/binary,
the new file gains it). No logic, signatures, or behavior change.
